+++ lib/nnq/routing/surveyor.rb
@@ -98,7 +98,7 @@ module NNQ
 
 
       def send_queue_drained?
-        @queues.each_value.all?(&:empty?)
+        @queues.each_value.all? { |q| q.empty? }
       end
 
 
@@ -126,7 +126,7 @@ module NNQ
           loop do
             body = queue.dequeue
             conn.send_message(body)
-            @engine.emit_verbose_monitor_event(:message_sent, body: body)
+            @engine.emit_verbose_msg_sent(body)
           rescue EOFError, IOError, Errno::EPIPE, Errno::ECONNRESET
             break
           end